Create A wolf spied a kid that had strayed from his herd, and pursued him. The kid, finding that he could not escape, waited until the wolf came up, and then assuming a cheerful tone, said, “I see clearly enough that you must be eaten, but I prefer to die as pleasantly as I can. Play me, therefore, a few notes on your pipe before I die.” The wolf, who was a musical turn and always carried his pipe with him, played a tune and the kid danced. Soon the noise of the pipe brought the sheepdogs to the spot. The wolf ran off saying, “this is what happens when people stray away from their profession. My business was to play the butcher, not the piper.  MORAL: Those who stray from their true business often lose the prize in hand.
A wolf spied a kid that had strayed from his herd, and pursued him. The kid, finding that he could not escape, waited until the wolf came up, and then assuming a cheerful tone, said, “I see clearly enough that you must be eaten, but I prefer to die as pleasantly as I can. Play me, therefore, a few notes on your pipe before I die.” The wolf, who was a musical turn and always carried his pipe with him, played a tune and the kid danced. Soon the noise of the pipe brought the sheepdogs to the spot. The wolf ran off saying, “this is what happens when people stray away from their profession. My business was to play the butcher, not the piper.

MORAL: Those who stray from their true business often lose the prize in hand.